Trying to simplify things.

+++ 2022/Day 7/include/solution.hpp
@@ -19,9 +19,9 @@ bool foundInString(std::string target, T substring);
 template <typename T>
 bool foundInVector(const std::vector<T> &vector, const T &element);
 
-int partOne(const std::string fileName);
+double partOne(const std::string fileName);
 
-int partTwo(const std::string fileName);
+double partTwo(const std::string fileName);
 
 enum class COMMAND {
     CHANGE_DIRECTORY,
@@ -37,7 +37,7 @@ class DeviceSystem {
 
         void loadTerminalOutput(std::string fileName);
 
-        int sumDirectoriesSmallerThan100KB();
+        double sumDirectoriesSmallerThan100KB();
 
         void printFiles();
